Eables printing of the instructor guide
This now allows printing of the instructor guide with `courseware print
guide`. This is the "new style" instructor guide that doesn't yet exist.
We'll create it by moving the long form prose that exists in the current
instructor notes to a new `guide` section below each slide.

Instructor notes will become the quick bullets or couple of sentences
that presentation notes should be, and the guide will be a PDF like the
student guide, but with prose attached to each slide explaining how to
teach to the concepts, and articulating the learning arc it fit into.