This cleans up how variants work
This unifies how variants work instead of special-casing it within the
`printer`  subclass. It means that the proper json will be looked at
anywhere we use it. This makes things like the validators work.

There is one exception. When we validate obsolete slides, we can't look
at only a single variant. We need to look at the conglomerate of **all**
variants in order to determine if a slide is not referenced in any of
them.

There's still a significant amount of cleanup and refactoring that I
want to do, but I'm going to hold off on that until we get a little more
experience building the variants.