Generate release notes for shared content
Now that content belonging to a course cannot be identified solely by
it's path, the release notes generator has to be smarter. Note that this
will catch a few false positives in the short term as we merge in the
handful of virtual courses we're keeping.

TRAINTECH-1236 #resolved #time 3h